Added notification system, added repository implemetations, improved controllers, added inquiery logic

+++ Web-shop/Controllers/ProductController.cs
@@ -8,6 +8,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using Web_shop.DataAccess.Data;
+using Web_shop.DataAccess.Repository;
 using Web_shop.Models;
 using Web_shop.Models.ViewModels;
 using Web_shop.Utility;
@@ -17,18 +18,20 @@ namespace Web_shop.Controllers
     [Authorize(Roles = WC.AdminRole)]
     public class ProductController : Controller
     {
-        private readonly ApplicationDbContext _dbContext;
+        private readonly IRepository<Product> _productRepository;
+        private readonly IRepository<Category> _categoryRepository;
         private readonly IWebHostEnvironment _webHostEnvironment;
 
-        public ProductController(ApplicationDbContext dbContext, IWebHostEnvironment eviroment)
+        public ProductController(IRepository<Product> productRepository, IRepository<Category> categoryRepository, IWebHostEnvironment eviroment)
         {
-            _dbContext = dbContext;
+            _productRepository = productRepository;
+            _categoryRepository = categoryRepository;
             _webHostEnvironment = eviroment;
         }
 
         public IActionResult Index()
         {
-            var productst = _dbContext.Products.Include(x => x.Category);
+            var productst = _productRepository.All.Include(x => x.Category);
             return View(productst);
         }
 
@@ -36,17 +39,21 @@ namespace Web_shop.Controllers
         {
             var productVM = new ProductVM()
             {
-                Categories = _dbContext.Categories.Select(c => new SelectListItem() { Text = c.Name, Value = c.Id.ToString() })
+                Categories = _categoryRepository.All.AsEnumerable().Select(c => new SelectListItem() { Text = c.Name, Value = c.Id.ToString() })
             };
 
             if(id is null)
             {
+                TempData[WC.ErrorNotification] = $"Product id was null";
+
                 return View(productVM);
             }
 
-            var product = _dbContext.Products.Find(id);
+            var product = _productRepository.Find(id.Value);
             if(product is null)
             {
+                TempData[WC.ErrorNotification] = $"Product with id equals {id.Value} not found";
+
                 return NotFound();
             }
 
@@ -69,34 +76,39 @@ namespace Web_shop.Controllers
             {
                 var product = productVM.Product;
                 //File logic
-                var oldProduct = _dbContext.Products.AsNoTracking().FirstOrDefault(p => p.Id == product.Id);
+                var oldProduct = _productRepository.All.AsNoTracking().FirstOrDefault(p => p.Id == product.Id);
+                product.ImageLink = oldProduct.ImageLink;
                 if (HttpContext.Request.Form.Files.Any())
                 {
                     var file = HttpContext.Request.Form.Files[0];
                     var newFileName = Guid.NewGuid().ToString() + Path.GetExtension(file.FileName);
                     var fileStream = System.IO.File.Create(Path.Combine(_webHostEnvironment.WebRootPath, WC.ImagePath, newFileName));
                     file.CopyTo(fileStream);
-
-                    var oldFileFullPath = Path.Combine(_webHostEnvironment.WebRootPath, WC.ImagePath, oldProduct.ImageLink);
-                    if (System.IO.File.Exists(oldFileFullPath))
-                    {
-                        System.IO.File.Delete(oldFileFullPath);
-                    }
-                                            
                     product.ImageLink = newFileName;
+
+                    if (oldProduct.ImageLink is not null)
+                    {
+                        var oldFileFullPath = Path.Combine(_webHostEnvironment.WebRootPath, WC.ImagePath, oldProduct.ImageLink);
+                        if (System.IO.File.Exists(oldFileFullPath))
+                        {
+                            System.IO.File.Delete(oldFileFullPath);
+                        }
+                    }
                 }
 
-                product.ImageLink = oldProduct.ImageLink;
-                _dbContext.Products.Update(product);
-                _dbContext.SaveChanges();
+                _productRepository.Update(product);
+                _productRepository.Save();
+                TempData[WC.SuccessNotification] = $"Product was successfully updated";
 
-                return RedirectToAction("Index");
+                return RedirectToAction(nameof(Index));
             }
                         
-            _dbContext.Products.Add(productVM.Product);
-            _dbContext.SaveChanges();
+            _productRepository.Add(productVM.Product);
+            _productRepository.Save();
+            TempData[WC.SuccessNotification] = $"Product was successfully created";
 
-            return RedirectToAction("Index");
+
+            return RedirectToAction(nameof(Index));
         }
         
         public IActionResult Delete(int id)
@@ -106,22 +118,29 @@ namespace Web_shop.Controllers
                 return NotFound();
             }
 
-            var product = _dbContext.Products.Include(p => p.Category).FirstOrDefault(p => p.Id == id);
+            var product = _productRepository.All.Include(p => p.Category).FirstOrDefault(p => p.Id == id);
             if (product is null)
             {
+                TempData[WC.ErrorNotification] = $"Product with id equals {id} not found";
+
                 return NotFound();
             }
 
-            var imagePath = Path.Combine(_webHostEnvironment.WebRootPath, WC.ImagePath, product.ImageLink);
-            if (System.IO.File.Exists(imagePath))
+            if (product.ImageLink is not null)
             {
-                System.IO.File.Delete(imagePath);
+                var imagePath = Path.Combine(_webHostEnvironment.WebRootPath, WC.ImagePath, product.ImageLink);
+                if (System.IO.File.Exists(imagePath))
+                {
+                    System.IO.File.Delete(imagePath);
+                }
             }
 
-            _dbContext.Products.Remove(product);
-            _dbContext.SaveChanges();
+            _productRepository.Remove(product);
+            _productRepository.Save();
+            TempData[WC.ErrorNotification] = $"Product with id equals {id} was successfully deleted";
 
-            return RedirectToAction("Index");
+
+            return RedirectToAction(nameof(Index));
         }
     }
 }
